requirements.txt
langchain==0.0.170
streamlit-chat
openai
streamlit
youtube-transcript-api
pytube
InstructorEmbedding
sentence-transformers
faiss-cpu
pikepdf
--upgrade streamlit